# Conflicts:
#	models/HomeModel.php
#	tests/WhishlistModelTest.php

+++ models/HomeModel.php
@@ -280,16 +280,19 @@ class HomeModel extends BaseModel
     }
     public function deleteWhishList($paged)
     {
-        $allWhishlist = $this->getWhishlist();
-        foreach ($allWhishlist as $value) {
-            $md5 = md5($value['id'] . "chuyen-de-web-2");
-            if ($md5 == $paged) {
-                $sql = "DELETE FROM whishlist WHERE id =  " . $value['id'];
-                $whishlist = $this->delete($sql);
-                return $whishlist;
+        if(!empty($paged)){
+            $allWhishlist = $this->getWhishlist();
+            foreach ($allWhishlist as $value) {
+                $md5 = md5($value['id'] . "chuyen-de-web-2");
+                if ($md5 == $paged && !is_bool($paged)) {
+                    $sql = "DELETE FROM whishlist WHERE id =  " . $value['id'];
+                    $whishlist = $this->delete($sql);
+                    return $whishlist;
+                }
             }
+        }else{
+            return false;
         }
-        return false;
     }
     // --------------------- Manufacture ------------------ //
     // Hien thi data bang manufactures:
